+++ src/App.jsx
@@ -75,6 +75,8 @@ const HeadToHeadContest = lazy(() => import('./pages/HeadToHeadContest'));
 const WeeklyFeatureVote = lazy(() => import('./pages/WeeklyFeatureVote'));
 const WeeklyReferralContest = lazy(() => import('./pages/WeeklyReferralContest'));
 const SharedWalletGroups = lazy(() => import('./pages/SharedWalletGroups'));
+const PrivacyPolicy = lazy(() => import('./pages/PrivacyPolicy'));
+const TermsOfService = lazy(() => import('./pages/TermsOfService'));
 const AIContentHub = lazy(() => import('./pages/AIContentHub'));
 const Store = lazy(() => import('./pages/Store'));
 const AIAgentsSettings = lazy(() => import('./pages/AIAgentsSettings'));
@@ -308,6 +310,8 @@ const AuthenticatedApp = () => {
       <Route path="/WeeklyFeatureVote" element={<LayoutWrapper currentPageName="WeeklyFeatureVote"><WeeklyFeatureVote /></LayoutWrapper>} />
       <Route path="/WeeklyReferralContest" element={<LayoutWrapper currentPageName="WeeklyReferralContest"><WeeklyReferralContest /></LayoutWrapper>} />
       <Route path="/SharedWalletGroups" element={<LayoutWrapper currentPageName="SharedWalletGroups"><SharedWalletGroups /></LayoutWrapper>} />
+      <Route path="/PrivacyPolicy" element={<PrivacyPolicy />} />
+      <Route path="/TermsOfService" element={<TermsOfService />} />
       <Route path="/AIContentHub" element={<LayoutWrapper currentPageName="AIContentHub"><AIContentHub /></LayoutWrapper>} />
       <Route path="/Store" element={<LayoutWrapper currentPageName="Store"><Store /></LayoutWrapper>} />
       <Route path="/AIAgentsSettings" element={<LayoutWrapper currentPageName="AIAgentsSettings"><AIAgentsSettings /></LayoutWrapper>} />